Fix "system-deploy delete" exception handle and information about deleting.
With this commit when 'software system-deploy delete' is called it should display the files deleted and some information about success or error
And the logs are now corrected and errors are cast as string and not showing incorrect messages.

Test Plan:

PASS: Deleted system-deploy with command 'software system-deploy delete' and check the information are shown correctly in the logs and on screen.
PASS: Build and deploy test-iso to SX and DX and check the command "live" on local environment.

Story: 2011720
Task: 54711

+++ software/software/software_controller.py
@@ -3072,16 +3072,26 @@ class PatchController(PatchService):
         return release_info
 
     def _remove_backup_files(self, bkp_info):
+        msg_info = ""
+        msg = ""
         for bkp_path, fail_warning_msg in bkp_info.items():
+            msg = "Deleted existing %s" % bkp_path
             if os.path.isdir(bkp_path):
                 shutil.rmtree(bkp_path)
-                LOG.info("Deleted existing %s", bkp_path)
+                LOG.info(msg)
             else:
                 try:
-                    os.remove(bkp_path)
-                    LOG.info("Deleted existing %s", bkp_path)
+                    if os.path.exists(bkp_path):
+                        os.remove(bkp_path)
+                        LOG.info(msg)
+                    else:
+                        msg = "There is no %s to be deleted. Skipped" % bkp_path
+                        LOG.info(msg)
                 except OSError as e:
-                    LOG.warning(fail_warning_msg + e)
+                    msg = fail_warning_msg + str(e)
+                    LOG.warning(fail_warning_msg + str(e))
+            msg_info += "%s\n" % msg
+        return msg_info
 
     def software_system_deploy_delete_api(self) -> dict:
         """
@@ -3109,7 +3119,7 @@ class PatchController(PatchService):
 
             # Deleting backup /etc/kubernetes/manifests.bkp, /etc/containerd/config.toml.bkp
             # and backup /opt/backups/k8s-control-plane
-            self._remove_backup_files(files_to_delete)
+            msg_info += self._remove_backup_files(files_to_delete)
         except Exception as e:
             msg = "Failed on deleting system-deploy"
             msg_error += msg + ".\n"
